requirements.txt
absl-py==2.1.0
astroid==3.3.8
cfgv==3.4.0
contourpy==1.3.1
coverage==7.6.10
cycler==0.12.1
dill==0.3.9
distlib==0.3.9
dnspython==2.7.0
filelock==3.17.0
fonttools==4.55.8
fsspec==2025.2.0
grpcio==1.70.0
identify==2.6.6
iniconfig==2.0.0
isort==6.0.0
Jinja2==3.1.5
kiwisolver==1.4.8
Markdown==3.7
MarkupSafe==3.0.2
matplotlib>=3.4.0
mccabe==0.7.0
mpmath==1.3.0
networkx==3.4.2
nodeenv==1.9.1
numpy>=1.21.0
packaging==24.2
pandas>=1.3.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
pre_commit==4.1.0
protobuf==5.29.3
psutil>=5.8.0
psycopg2-binary==2.9.10
pylint==3.3.4
pymongo==4.11
pyparsing==3.2.1
pytest==8.3.4
pytest-cov==6.0.0
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML>=5.4.0
seaborn>=0.11.0
setuptools==75.8.0
six==1.17.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tomlkit==0.13.2
torch>=2.0.0
torchvision==0.21.0
tqdm>=4.62.0
typing_extensions==4.12.2
tzdata==2025.1
virtualenv==20.29.1
Werkzeug==3.1.3
mlflow>=2.0.0
numba>=0.54.0
scikit-learn>=0.24.0
